+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Fix dbapi instrumentation with MySQL
+  ([#1183](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1183))
 - Add uninstrument test for sqlalchemy
   ([#1471](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/1471))
 - `opentelemetry-instrumentation-tortoiseorm` Initial release